Physics bug fix
Physics had access violation because the rigidbody was bein set to null ptr, which it does when no rigidbody is found, actually there should be a safety net for that so it doesn't just stop the program and throws an error cause of a missing rigidbody which is allowed to be missing. Also moved the physics calculation from the math class to the physics class as it does make more sense for it to be stored there altough the physics class becomes a bit bigger because of it. But files that include the math file should not have access to the collision functions, and now they are safely made private and inacessable within physics also less code needs to be compiled from math when including it wich will help to keep the compiel time low.
Another small thing, the component class has a (reflective) toString method so what ever child class you have of component that class name will be printed when calling it's toString, very nice
