+++ SFML_Tutorial/Scene01.txt
@@ -1,1 +1,1 @@
-945.000000,830.000000
+581.000000,850.000000
